feat: moved Dashboard component to Home page directory
