+++ src/components/Pages/Home/Dashboard/BarContainer/Bar/index.tsx
