+++ src/components/Pages/Home/Dashboard/index.tsx
@@ -5,7 +5,7 @@ import {
 	SidePanel,
 	MainPanel,
 	BottomPanel
-} from 'src/components/Dashboard/Panels'
+} from 'src/components/Pages/Home/Dashboard/Panels'
 
 const DashBoard: React.FC = () => {
 	return (
