+++ src/components/Pages/Home/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 // components
-import Dashboard from 'src/components/Dashboard'
+import Dashboard from 'src/components/Pages/Home/Dashboard'
 
 const Home: React.FC = () => {
 	return (
